refactor: split chat screen components
